refactored collections name to features, import at init and fixed arraynotinstalled abc missing functions
