+++ merlin/features/__init__.py
@@ -13,3 +13,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+from merlin.features.features import Feature, Features
+
+__all__ = [
+    "Feature",
+    "Features",
+]
