+++ merlin/features/array/numpy.py
@@ -14,7 +14,7 @@
 # limitations under the License.
 #
 from merlin.features.array.base import ArrayPackageNotInstalled, MerlinArray
-from merlin.features.array.compat import numpy
+from merlin.features.compat import numpy
 
 if not numpy:
