+++ tests/unit/features/array/test_cudf.py
@@ -16,11 +16,12 @@
 import pytest
 
 from merlin.core.dispatch import make_df
-from merlin.features.array.compat import cudf, cupy, numpy, pandas, tensorflow
 from merlin.features.array.cudf import MerlinCudfArray
+from merlin.features.compat import cudf, cupy, numpy, pandas, tensorflow
 from merlin.features.df import VirtualDataFrame
 
-pytest.importorskip("cudf")
+if cudf is None:
+    pytest.skip("cudf is not installed.", allow_module_level=True)
 
 
 @pytest.mark.skipif(numpy is None, reason="Numpy is not defined")
@@ -50,7 +51,7 @@ def test_cudf_series_to_merlin_cudf_array():
     assert (cupy.asnumpy(merlin_cudf_array.array) == cudf_series.to_numpy()).all()
 
 
-@pytest.mark.skipif(tensorflow is None, reason="Tensorflow is not defined")
+@pytest.mark.skipif(tensorflow is None, reason="tensorflow is not installed")
 def test_tf_tensor_to_merlin_cudf_array():
     tf_tensor = tensorflow.random.uniform((10,))
     merlin_cudf_array = MerlinCudfArray(tf_tensor)
