+++ tests/unit/features/array/test_tensorflow.py
@@ -15,11 +15,12 @@
 #
 import pytest
 
-from merlin.features.array.compat import cudf, cupy, numpy, pandas, tensorflow
 from merlin.features.array.tensorflow import MerlinTensorflowArray
+from merlin.features.compat import cudf, cupy, numpy, pandas, tensorflow
 from merlin.features.df import VirtualDataFrame
 
-pytest.importorskip("tensorflow")
+if tensorflow is None:
+    pytest.skip("tensorflow is not installed.", allow_module_level=True)
 
 
 @pytest.mark.skipif(numpy is None, reason="Numpy is not defined")
@@ -49,7 +50,7 @@ def test_cudf_series_to_merlin_tf_array():
     assert (merlin_tf_array.array.numpy() == cudf_series.to_numpy()).all()
 
 
-@pytest.mark.skipif(tensorflow is None, reason="Tensorflow is not defined")
+@pytest.mark.skipif(tensorflow is None, reason="tensorflow is not installed")
 def test_tf_tensor_to_merlin_tf_array():
     tf_tensor = tensorflow.random.uniform((10,))
     merlin_tf_array = MerlinTensorflowArray(tf_tensor)
